+++ CHANGELOG.md
@@ -8,6 +8,13 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Changed
+
+- Used unique `deviceStatus.id` for `_key` property of host agent ->
+  configuration relationships.
+- Used unique `deviceStatus.id` for `_key` property of device -> managed
+  application relationships.
+
 ## 3.0.0 - 2020-03-18
 
 ### Updated